pyproject.toml
[tool.poetry]
name = "scrapy_basics"
version = "0.1.0"
description = "Learn Scrapy"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Scrapy = "^2.11.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
